Use TaskOptions in task update
Using TaskOptions allows the user to specify an inbound
hosted_zone_name, or 'auto', keeping this out of the task model.

As this is a PATCH, as a result of breaking Task down, there is no
longer an 'id' field required in the request body The task ID is already
specified in the path.